backend/requirements.txt
# ── 웹 프레임워크 ──
fastapi==0.115.5
uvicorn[standard]==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.1.0
apscheduler==3.10.4
loguru==0.7.2
httpx==0.27.2

# ── 크롤링 ──
requests
beautifulsoup4==4.12.3
lxml==5.2.2
playwright==1.44.0

# ── 데이터 처리 ──
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0

# ── ML ──
xgboost==2.0.3
scikit-learn==1.4.2
optuna==3.6.1
joblib==1.4.2

# ── DL ──
torch==2.3.1
tqdm==4.66.4

# ── AI ──
openai

# ── 시각화 ──
matplotlib==3.9.0
jupyter==1.0.0
